requirements.txt
future==1.0.0
pymongo==4.12.1
python-telegram-bot==22.0
